Wait longer before retrying and ignore invalid bytes
When an error occurs, simply re-sending the failed command doesn't
always work. Sometimes we appear to receive some stale data from the
previous attempt, despite the fact that the input queue is purged before
retrying.

To improve the chances of re-synchronizing on the start of the next
packet, wait a bit longer before re-trying and also ignore all received
bytes until an ACK byte is encountered. This isn't entirely reliable,
because the payload data isn't escaped and can also contain ACK bytes,
but it's better than nothing.

Below is a real-life example of such a case.

First, we hit a timeout:

  [8.122494] INFO: Write: size=2, data=E742
  [8.122663] INFO: Write: size=8, data=708F020080000000
  [11.130287] INFO: Read: size=0, data=
  [11.130308] ERROR: Failed to receive the answer. [in src/mares_iconhd.c:213 (mares_iconhd_packet)]

Next, we retry a few times:

  [11.130321] INFO: Sleep: value=100
  [11.230536] INFO: Purge: direction=1
  [11.231117] INFO: Write: size=2, data=E742
  [11.240494] INFO: Write: size=8, data=708F020080000000
  [11.415317] INFO: Read: size=20, data=0000000035004F000C00000037004F0000000000
  [11.415331] ERROR: Unexpected answer byte. [in src/mares_iconhd.c:219 (mares_iconhd_packet)]
  [11.415339] INFO: Sleep: value=100
  [11.515889] INFO: Purge: direction=1
  [11.517474] INFO: Write: size=2, data=E742
  [11.517628] INFO: Write: size=8, data=708F020080000000
  [11.517798] INFO: Read: size=13, data=1300000036004F000A00000037
  [11.517803] ERROR: Unexpected answer byte. [in src/mares_iconhd.c:219 (mares_iconhd_packet)]

That didn't really work, but the interesting part is that we did receive
some data now. We also see the same data appear in the next retry:

  [11.517811] INFO: Sleep: value=100
  [11.618398] INFO: Purge: direction=1
  [11.618760] INFO: Write: size=2, data=E742
  [11.619100] INFO: Write: size=8, data=708F020080000000
  [11.619341] INFO: Read: size=1, data=AA
  [11.619349] INFO: Read: size=20, data=0000000035004F000C00000037004F0000000000
  [11.619355] INFO: Read: size=20, data=37004F001E00000036004F001000000037004F00
  [11.619360] INFO: Read: size=20, data=2C00000037004F001200000034004F0033000000
  [11.619366] INFO: Read: size=20, data=37004F000000000037004F001E00000036004F00
  [11.630104] INFO: Read: size=12, data=1300000036004F000A000000
  [11.630122] INFO: Read: size=20, data=37004F002000000038004F001D00000038004F00
  [11.630131] INFO: Read: size=17, data=2300000039004F000600000037004F00EA

This time, the retry seems to have worked and we move on to the next
request:

  [11.635160] INFO: Write: size=2, data=E742
  [11.635359] INFO: Write: size=8, data=F08E020080000000
  [11.687620] INFO: Read: size=1, data=AA
  [11.708476] INFO: Read: size=20, data=0000000035004F000C00000037004F0000000000
  [11.719214] INFO: Read: size=20, data=37004F001E00000036004F001000000037004F00
  [11.719231] INFO: Read: size=20, data=2C00000037004F001200000034004F0033000000
  [11.719240] INFO: Read: size=20, data=37004F000000000037004F001E00000036004F00
  [11.740569] INFO: Read: size=13, data=1300000036004F000A00000037
  [11.740585] INFO: Read: size=20, data=004F002000000038004F001D00000038004F0023
  [11.740592] INFO: Read: size=20, data=00000039004F000600000037004F00EAAA090010

This appears to have worked, but upon closer inspection you can see that
the very last packet already contains the start of the response to the
next request (AA090010). But we haven't send the next request yet!

That means the response we received here is actually the response to one
of the earlier (failed) request and this start of the next packet is our
real response. So we're getting completely out of sync! Unfortunately
the Mares protocol doesn't use checksums or sequence numbers, so we
can't really detect this problem.