Add support for the Cressi BLE protocol variant
The BLE variant of the Cressi protocol uses a completely different set
of commands compared to the existing usb-serial protocol. Most of the
packet framing bytes have been removed as well. But the most surprising
change is the fact that there is no variant of the CMD_VERSION
available. The corresponding information must be obtained by reading
some secondary characteristics instead.

Co-authored-by: Greg McLaughlin <support@moremobilesoftware.com>
Co-authored-by: Sven Knoch <info@divinglog.de>