Add the device id and logbook entry
Previously, in commit 57cd11fffe4c44d6323a4d5b2bfd04e9e0f5a7a4, the part
of the logbook entry containing the dive mode was already inserted back
into the dive data, because the dive mode is needed to be able to parse
the dive.

Unfortunately, the newer models like the Cressi Nepto store even more
important information in the logbook entry and the dive computer id
data. Therefore, the data format is changed once more to preserve both
pieces of data. The parser is updated to handle this new format.

This change is not strictly backwards compatible, but if needed existing
data can be converted into the new format.