Add CCR support for the Excursion
There is a minor update to the dataformat to support CCR dives.

For the current firmware, the setpoint change event was originally
defined as a 4 byte event, but the new implementation changed this to
only 2 bytes. Since the current firmware doesn't support CCR dives yet,
the 4 bytes variant is not supposed to appear in the sample data and
should be ignored.