Use a layout descriptor with a generic decode function
Replace hardcoded constants with a layout descriptor. This reduces the
amount of model specific conditions, and makes it easier to add support
for new models.

The Cressi data format uses nibbles (4 bits) as the smallest unit of
storage. Larger values are stored as one or more nibble, typically with
a BCD (binary coded decimal) encoding applied. To decode such variable
sized numbers, a generic function is added. The offsets in the layout
descriptor are also specified in nibbles instead of bytes.