Add BLE support for the Seac Tablet
The Seac Tablet supports both usb-serial and BLE communication. Both
transports share the same high-level communication protocol. Only the
packetization is different.

The BLE communication is based on the typical UART-over-BLE protocol,
featuring the following service and Rx/Tx characteristics:

  Service: 84968ffe-d26d-478a-b953-5010bcf58bca
  Characteristics:
    - Rx/Tx: 43c620c2-1b09-4951-bc1e-9c75298cddeb (read, write)

Remarkably, the Tx characteristic does not support notifications or
indications and must be read instead.

The bluetooth device name is "TabletXXXXXX" where XXXXXX is the serial
number of the dive computer.