Fix the date/time parsing
The Symbios stores the date/time value as a UTC timestamp. Since there
is no timezone information present, I incorrectly assumed the timestamp
actually represented local time. Therefore I used the gmtime function to
perform a timezone independent conversion.

Fixed by using the localtime function instead.

Note: this approach makes the conversion dependent on the current
timezone of the host system, but there is no guarantee this timezone is
indeed the correct timezone in which the dive was recorded!