Add support for new OSTC 4/5 events
In recent OSTC 4/5 firmware versions, the data format gained support for
several new events:

 - Compass heading (v1.6.9)
 - GNSS position (v1.7.0)
 - Scrubber state (v1.7.2)

Since events are located before the extended sample data, the presence
of unsupported events is problematic. That's because the size of the
event isn't known and thus the corresponding bytes can't be skipped
correctly. Instead, those bytes will get interpreted as the next event
or as the extended sample data, resulting in some bogus values. This
problem is limited to the current sample only, because the total length
of the entire sample is available in the data.