Make the record validation non-fatal during the download
When the dive computer battery voltage is too low, the external memory
read/write operations are not performed, which can result in corrupt
records with an incorrect checksum. The current code considers such
checksum errors as a fatal error and aborts the download. This prevents
the user from downloading any older dives.

However, during the download the contents of the header isn't really
needed, except for two cases:

 - To determine the length of the most recent dive and calculate the
   end-of-profile pointer.

 - To detect and discard padding bytes at the end of each dive.

In both cases, there is a reasonable fallback possible whenever the
header checksum is invalid and the contents of the header can't be
trusted.

In the parser, checksum errors are still treated as fatal errors.