Optimize the download using the dive number as fingerprint
Previously, the date/time information in the dive header was used as the
dive fingerprint. In order to decide which dives need to be downloaded,
this choice requires downloading the header of each dive during the
first pass of the download. When the entire dive is downloaded in the
second pass, these dive headers are downloaded a second time.

Using the dive number as the fingerprint can eliminate this first
download of the dive headers. That's possible because the range of
available dive numbers is communicated earlier by the dive computer and
thus the fingerprint (dive number of the most recent dive) can be
applied immediately.