Fix the Github actions on mac
After Github updated the macos-latest image to macOS 15 (with Xcode 16),
linking fails with the error:

  ld: Shared cache eligible dylib cannot link to ineligible dylib
  '/opt/homebrew/opt/libusb/lib/libusb-1.0.0.dylib'.  Remove link to
  ineligible dylib, fix its eligibility, or opt out of the shared cache
  using the build setting 'LD_SHARED_CACHE_ELIGIBLE=NO' (or linker flag
  '-not_for_dyld_shared_cache')

The earlier macOS 14 image (with Xcode 15.4) already issued a warning:

  ld: warning: invalid -install_name
  (/opt/homebrew/opt/libusb/lib/libusb-1.0.0.dylib) in dependent dylib
  (/opt/homebrew/Cellar/libusb/1.0.29/lib/libusb-1.0.dylib).
  Dylibs/frameworks which might go in dyld shared cache cannot link with
  dylibs that won't be in the shared cache

This problem appears to be caused by running the configure script with
the --prefix=/usr option. Remove this option and use the default
/usr/local prefix instead of enabling some obscure linker option.

For consistency, the prefix option is removed from all build targets.