Replace the hardware descriptor with the model number
The different bits in the hardware descriptor describe the available
hardware features:

  0  Rechargeable battery with battery management chip
  1  Ambient sensor
  2  Analog inputs and S8 digital
  3  Digital optical input
  4  BLE module
  5  32bit dual core
  6  Low voltage core
  7  External flash memory with block write support

Since different models can share the same hardware features, the
hardware descriptor can no longer be used to reliably identify the exact
model. Trying to maintain this relationship anyway, only results in many
duplicated entries in the list of supported devices with the same name
but a different model number.

For this reason, the hwOS firmware even masks off the upper 2 bits of
the hardware descriptor before sending the value.

Use the model descriptor as the internal model number instead. All the
hwOS models, except for the OSTC 4 and 5, report a zero value here.